requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
idna==3.11
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pip==25.3
polars==1.35.1
polars-runtime-32==1.35.1
psutil==7.1.3
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyyaml==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing-extensions==4.15.0
ultralytics @ git+https://github.com/ultralytics/ultralytics.git@667f306841ee01c926d4efeea9999b8e73409b38
ultralytics-thop==2.0.18
urllib3==2.5.0
